Corrected database name that wasn't corresponding to our database (see DatabaseHelper and UserTable) + updated UserNote + updated fragment_profil.xml
 Changes to be committed:
	modified:   app/src/main/java/bam/bam/bam/dataBDD/UserDAO.java
	modified:   app/src/main/java/bam/bam/bam/dataBDD/UserTable.java
	modified:   app/src/main/java/bam/bam/bam/modeles/User.java
	modified:   app/src/main/java/bam/bam/bam/modeles/UserNote.java
	modified:   app/src/main/java/bam/bam/bam/views/fragment/ProfilFragment.java
	modified:   app/src/main/java/bam/bam/globalDisplay/database/DatabaseHelper.java
	modified:   app/src/main/res/layout/fragment_profil.xml

+++ app/src/main/java/bam/bam/bam/modeles/User.java
@@ -47,7 +47,7 @@ public class User {
      */
     private String photo_data ;
 
-    public User(int id, String user_pseudo, String user_device_id, String user_phone_number,String photo_data, int note, String status, int nbn ) {
+    public User(int id, String user_pseudo, String user_device_id, String user_phone_number,String photo_data, float note, String status, int nbn ) {
 
         this.id = id;
         this.photo_data  = photo_data ;
